--- a/src/HaskellWorks/Data/BalancedParens/Internal/Broadword/FindClose/Word8.hs
+++ b/src/HaskellWorks/Data/BalancedParens/Internal/Broadword/FindClose/Word8.hs
@@ -48,6 +48,6 @@ findClose :: Word8 -> Count -> Maybe Count
 findClose v p = if p > 0
   then if closeAt v p
     then Just p
-    else let q = W8.findUnmatchedCloseFar 0 p v in Just (q + 1)
+    else Just (W8.findUnmatchedCloseFar 0 p v + 1)
   else Just (W8.findUnmatchedCloseFar 1 p v)
 {-# INLINE findClose #-}