For a small, but respectable performance gain.

Before:

```
benchmarking Vector/FindClose 2-bit/Broadword
time                 19.30 ns   (19.05 ns .. 19.65 ns)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 19.28 ns   (19.14 ns .. 19.46 ns)
std dev              537.2 ps   (387.1 ps .. 688.1 ps)
variance introduced by outliers: 45% (moderately inflated)

benchmarking Vector/FindClose 4-bit/Broadword
time                 19.51 ns   (19.15 ns .. 20.05 ns)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 19.62 ns   (19.39 ns .. 19.99 ns)
std dev              983.9 ps   (661.5 ps .. 1.325 ns)
variance introduced by outliers: 73% (severely inflated)

benchmarking Vector/FindClose 8-bit/Broadword
time                 19.28 ns   (19.16 ns .. 19.39 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 19.23 ns   (19.15 ns .. 19.36 ns)
std dev              334.4 ps   (251.4 ps .. 467.0 ps)
variance introduced by outliers: 24% (moderately inflated)

benchmarking Vector/FindClose 16-bit/Broadword
time                 19.32 ns   (19.24 ns .. 19.40 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 19.27 ns   (19.22 ns .. 19.33 ns)
std dev              196.4 ps   (136.6 ps .. 262.5 ps)
variance introduced by outliers: 10% (moderately inflated)

benchmarking Vector/FindClose 32-bit/Broadword
time                 19.27 ns   (19.20 ns .. 19.34 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 19.23 ns   (19.15 ns .. 19.31 ns)
std dev              262.3 ps   (229.0 ps .. 312.0 ps)
variance introduced by outliers: 17% (moderately inflated)
```

After:

```
benchmarking Vector/FindClose 2-bit/Broadword
time                 18.77 ns   (18.63 ns .. 18.91 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 18.68 ns   (18.61 ns .. 18.76 ns)
std dev              262.2 ps   (207.4 ps .. 317.9 ps)
variance introduced by outliers: 17% (moderately inflated)

benchmarking Vector/FindClose 4-bit/Broadword
time                 18.56 ns   (18.50 ns .. 18.63 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 18.57 ns   (18.52 ns .. 18.66 ns)
std dev              210.0 ps   (131.2 ps .. 392.3 ps)
variance introduced by outliers: 12% (moderately inflated)

benchmarking Vector/FindClose 8-bit/Broadword
time                 18.61 ns   (18.54 ns .. 18.68 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 18.64 ns   (18.59 ns .. 18.71 ns)
std dev              192.5 ps   (142.2 ps .. 261.8 ps)
variance introduced by outliers: 10% (moderately inflated)

benchmarking Vector/FindClose 16-bit/Broadword
time                 18.67 ns   (18.57 ns .. 18.77 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 18.64 ns   (18.59 ns .. 18.70 ns)
std dev              177.3 ps   (142.8 ps .. 235.6 ps)

benchmarking Vector/FindClose 32-bit/Broadword
time                 18.61 ns   (18.55 ns .. 18.68 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 18.65 ns   (18.58 ns .. 18.75 ns)
std dev              255.7 ps   (174.0 ps .. 382.0 ps)
variance introduced by outliers: 17% (moderately inflated)

benchmarking Vector/FindClose 64-bit/Broadword
time                 18.61 ns   (18.53 ns .. 18.79 ns)
                     0.999 R²   (0.995 R² .. 1.000 R²)
mean                 18.77 ns   (18.59 ns .. 19.42 ns)
std dev              1.049 ns   (188.6 ps .. 2.175 ns)
variance introduced by outliers: 77% (severely inflated)
```